+++ test/memento/multi_test.clj
@@ -4,7 +4,7 @@
             [memento.config :as mc]
             [clojure.test :refer :all]))
 
-(def inf-cache {mc/type mc/guava})
+(def inf-cache {mc/type mc/caffeine})
 
 (defn as-map [cache] (reduce-kv #(assoc %1 (:args %2) %3) {} (b/as-map cache)))