+++ CHANGELOG.md
@@ -1,4 +1,19 @@
 # Changelog
+
+## 1.0.x
+- remove Guava and replace with Caffeine
+- rewrite the readme
+- mark Guava namespaces for deprecation
+
+## 0.9.36 (2022-03-01)
+- add `memento.core/defmemo`
+
+## 0.9.35 (2022-02-22)
+- add `memento.core/key-fn*` mount setting
+
+## 0.9.34 (2022-02-16)
+- add support for using meta 
+
 ## 0.9.3 (2021-05-08)
 ### Features added
 - `memento.core/if-config`